requirements.txt
streamlit
pandas
matplotlib
python-binance
python-telegram-bot==13.15
ta
python-dotenv